web/tests/requirements.txt
pytest==8.3.2
pytest-django==4.8.0
beautifulsoup4==4.12.3
freezegun==1.5.1
